+++ manual/release-notes.rst
@@ -33,7 +33,7 @@ more detail.
 
   - Bug fixes
 
-    - Set `is_different` flag in `QPDFFormFieldObjectHelper::getTopLevelField` to
+    - Set ``is_different`` flag in ``QPDFFormFieldObjectHelper::getTopLevelField`` to
       false if the field is a top-level field. Previously the flag was only set
       if the field is a top-level field.
 
@@ -58,7 +58,15 @@ more detail.
     - Add new ``Buffer`` methods ``move``, ``view``, ``data``, ``size`` and
       ``empty``. The new methods present the ``Buffer`` as a ``char`` (rather
       than ``unsigned char``) container and facilitate the efficient moving
-      of its content into a `std::string``.
+      of its content into a ``std::string``.
+
+    - Add various new functions in the ``qpdf::`global`` namespace to access
+      and set/modify global settings and limits. See :ref:`global-options`
+      and header file ``qpdf/global.hh`` for further detail.
+
+    - Add new C-API functions ``qpdf_global_get_uint32`` and
+      ``qpdf_global_set_uint32`` to access and set/modify various global
+      settings and limits.
 
   - Build fixes
 
@@ -73,6 +81,9 @@ more detail.
 
     - Option :qpdf:ref:`--check` now includes additional basic checks of the
       AcroForm, Dests, Outlines, and PageLabels structures.
+
+    - Add new option :qpdf:ref:`--global` to set or modify various global
+      options and limits. See :ref:`global-options` for further detail.
 
     - Fix completion scripts and handling to avoid leaking arguments
       into the environment during completion and to correctly handle